requirements.txt
streamlit
google-genai
pandas
pdfplumber
openpyxl
python-docx
